feat: add the "create-rpm-dir" input that specifies if a rpm directory must be created
In our case, the "rpm" folder should not be created when we upload new rpm to devarch. For instance, \\devarch\archives\90\32\Installers\linux\Release4 contains all the RPM in the Release4 folder.
That is why I suggest the "create-rpm-directory" input that will be equals to true by default so it won't impact the other users.